+++ lib/load.php
@@ -125,6 +125,7 @@ require __DIR__ . '/compat/wordpress-7.1/media.php';
 require __DIR__ . '/compat/wordpress-7.1/preload.php';
 require __DIR__ . '/compat/wordpress-7.1/classic-block.php';
 require __DIR__ . '/compat/wordpress-7.1/icons.php';
+require __DIR__ . '/compat/wordpress-7.1/font-face.php';
 
 // Experimental features.
 require __DIR__ . '/experimental/omnibar/load.php';